requirements.txt
langchain
langchain-community
langchain-huggingface
mysql-connector-python
sqlalchemy
python-dotenv
streamlit
huggingface_hub